PostgreSQL restore from file encoding fix
Encoding was not cleanly set on each loop run.

Encoding was set, it was never set again and the first set is often
NONE and that ended in not creating any databases

First check if global encoding setting is set, set this.
Then check if we have per file encoding, set this, else set the global
default one.