dry run flag update for drop restore, wal compression update
The test run flag is now -n (dry run) in the drop restore postgresql
script.

In the WAL copy script it is now possible to define the compression
software and the default one changed from bzip2 to lzop to have less
cpu/disk io usage but still good compression. [-x <software>]
It is also possible to move the file post the compression to another
folder (eg to a NFS mounted one) [-m <path>]