Hot fix for Postgresql 9.6.2 dump schema ACL issue
In PostgreSQL 9.6.2 the dump command does not dump the default schema
ACLs anymore (bug?). So once the data is restored the DB owner cannot
see the tables.
The script checks if in the dump file the ACL entry for the public
schema is missing and then creates the default entries.

See more info:
http://www.postgresql-archive.org/intentional-or-oversight-pg-dump-c-does-not-restore-default-priviliges-on-schema-public-td5943782.html