PG dump log location check, PG dump restore role fix
PG dump log location fix for if the file or folder is not writeable.
Fallback to current folder or to home directory of user running script.

PG dump restore now adds missing --role command if the owner is set on
the command line.
Also adds new -s flag for only importing schema and not data.

Note that only import data and no schema is not supported and a normal
import should be run.
Nor is importing only a view tables. Those should be run from the
command line directly.
Run script with -n command for getting the basic pg_restore call