Borg backup update with init and check flags
Borg backup script does no longer automatically check if the repository
exists and also no longer auto inits the repository.

The following new options and settings are available
Settings:
FORCE_CHECK="true|false"
If set to true it will force a repository check (legacy flow)

Options:
-C: check if repository exists, if OK continue to backup, else exit
-E: after check exit, even if repository exists
-I: initialze a new repository. Must be run prior to running backups

A new file will be created in the config folder called
"borg.backup.init" with the timestamp when the repository was created.

If this file does not exist a warning will be displayed and the backup
will abort.

For old repsoitorys a one time -C -E run is needed.

Not if the folder/repository changes the general borg created command
will abort